% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_circle_ggplot.R
\name{annotate_circle_ggplot}
\alias{annotate_circle_ggplot}
\title{Function to help draw a cirle in ggplot2.}
\usage{
annotate_circle_ggplot(radius, x_center, y_center, color = "black",
  fill = NA, ...)
}
\arguments{
\item{radius}{is the radius}

\item{x_center}{is the x0 coordiante}

\item{y_center}{is the y0 coordiante}

\item{color}{is the color of the line}

\item{fill}{is whether to fill with a color}
}
\value{
A ggplot2 line object that is a circle.
}
\description{
I found it on the internet, I need to give credit at some point. But this
helps to create circles at a specified radius from a cartesian point.
}
\details{
This function is not exported.
}
