% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{cart_to_pol}
\alias{cart_to_pol}
\title{Converts cartesian coordinates to polar coordinates.}
\usage{
cart_to_pol(x, y)
}
\arguments{
\item{x}{is the "x" value on the cartesian coordinate system}

\item{y}{is the "y" value on the cartesian coordinate system}
}
\value{
list with converted "radius" and "theta" coordinates
}
\description{
Converts cartesian coordinates to polar coordinates.
}
