% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dartboard.R
\name{create_dartboard}
\alias{create_dartboard}
\title{Create a dartboard with Standard measurments.}
\usage{
create_dartboard()
}
\value{
A list object containing data.table objects describing the dimesions of
        the dartboard. These are stored in data.tables called bed_thetas, bed_values,
        bull_values, outer_ring_values
}
\description{
This function returns a list containing data.table objects describing the dimesions of
        the dartboard. These are stored in data.tables called bed_thetas, bed_values,
        bull_values, outer_ring_values.
}
