% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dartboard.R
\name{draw_dartboard}
\alias{draw_dartboard}
\title{Draw the dartboard object, db, to actual scale using ggplot2.}
\usage{
draw_dartboard(db)
}
\arguments{
\item{db}{a dartboard object created by dartboard::create_dartboard()}
}
\value{
A a ggplot2 object drawing of the dartboard.
}
\description{
Draw the dartboard object, db, to actual scale using ggplot2.
}
