% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{pol_to_cart}
\alias{pol_to_cart}
\title{Converts polar coordinates to cartesian coordinates.}
\usage{
pol_to_cart(radius, theta)
}
\arguments{
\item{radius}{radius of the vector}

\item{theta}{angles of the vector}
}
\value{
list with converted "x" and "y" coordinates
}
\description{
Converts polar coordinates to cartesian coordinates.
}
